Приложение accounts переименовано на profile модель User теперь называется Profile изменены ссылки на модель User во всех приложениях на Profile
