--- echo/accounts/apps.py
@@ -1,11 +1,0 @@
-from django.apps import AppConfig
-
-
-class AccountsConfig(AppConfig):
-    name = 'accounts'
-
-    def ready(self):
-        import accounts.signals
-
-
-
