--- echo/accounts/models.py
@@ -1,29 +1,0 @@
-from PIL import Image
-
-from django.db import models
-from django.contrib.auth.models import User
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(User, null=True, on_delete=models.CASCADE)
-    bio = models.TextField(null=True, blank=True)
-    image = models.ImageField(default='default.jpg', upload_to='profile_pics')
-    vk = models.CharField(max_length=50, null=True, blank=True)
-    telegram = models.CharField(max_length=50, null=True, blank=True)
-    whatsup = models.CharField(max_length=50, null=True, blank=True)
-    facebook = models.CharField(max_length=50, null=True, blank=True)
-    twitter = models.CharField(max_length=50, null=True, blank=True)
-    instagram = models.CharField(max_length=50, null=True, blank=True)
-
-    def __str__(self):
-        return f'{self.user.username} Profile'
-
-    def save(self, *args, **kwargs):
-        super(Profile, self).save(*args, **kwargs)
-
-        img = Image.open(self.image.path)
-
-        if img.height > 300 or img.width > 300:
-            output_size = (300, 300)
-            img.thumbnail(output_size)
-            img.save(self.image.path)
