+++ echo/posts/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
 from .models import Post, Category
-from sign.models import User
 
 
 class CategoryAdmin(admin.ModelAdmin):
@@ -11,4 +10,3 @@ class CategoryAdmin(admin.ModelAdmin):
 # Register your models here.
 admin.site.register(Post)
 admin.site.register(Category, CategoryAdmin)
-admin.site.register(User)
