+++ echo/profile/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class ProfileConfig(AppConfig):
+    name = 'profile'
+
+    def ready(self):
+        import profile.signals
